Model Validator Bug Fix
There was a condition where both teams could have been predicted as having won the match when the predicted probability for both teams was 0.5. This update fixes it by giving expected win priority to blue side for purposes of model validation (e.g. if both teams have win perc of 0.5, blue side is expected to win)